move LocationManagerClass inside CameraManager
