+++ camera/CameraManager.swift
@@ -12,6 +12,7 @@ import Photos
 import ImageIO
 import MobileCoreServices
 import Photos
+import CoreLocation
 
 public enum CameraState {
     case ready, accessDenied, noDeviceFound, notDetermined
@@ -1019,6 +1020,40 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
                 DispatchQueue.main.async() {
                     self._flipCameraTransitionView()
                 }
+            }
+        }
+    }
+
+    fileprivate class LocationManager: NSObject, CLLocationManagerDelegate {
+        var locationManager = CLLocationManager()
+        var latestLocation: CLLocation?
+
+        override init() {
+            super.init()
+            locationManager.delegate = self
+            locationManager.desiredAccuracy = kCLLocationAccuracyBest
+            locationManager.requestWhenInUseAuthorization()
+        }
+
+        func startUpdatingLocation() {
+            locationManager.startUpdatingLocation()
+        }
+
+        func stopUpdatingLocation() {
+            locationManager.stopUpdatingLocation()
+        }
+
+        // MARK: - CLLocationManagerDelegate
+        func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+            // Pick the location with best (= smallest value) horizontal accuracy
+            latestLocation = locations.sorted { $0.horizontalAccuracy < $1.horizontalAccuracy }.first
+        }
+
+        func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+            if status == .authorizedAlways || status == .authorizedWhenInUse {
+                locationManager.startUpdatingLocation()
+            } else {
+                locationManager.stopUpdatingLocation()
             }
         }
     }
