--- camera/LocationManager.swift
@@ -1,38 +1,0 @@
-import Foundation
-import CoreLocation
-
-class LocationManager: NSObject, CLLocationManagerDelegate {
-    var locationManager = CLLocationManager()
-    var latestLocation: CLLocation?
-    
-    override init() {
-        super.init()
-        locationManager.delegate = self
-        locationManager.desiredAccuracy = kCLLocationAccuracyBest
-        locationManager.requestWhenInUseAuthorization()
-    }
-    
-    func startUpdatingLocation() {
-        locationManager.startUpdatingLocation()
-    }
-    
-    func stopUpdatingLocation() {
-        locationManager.stopUpdatingLocation()
-    }
-    
-    // MARK: - CLLocationManagerDelegate
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        // Pick the location with best (= smallest value) horizontal accuracy
-        latestLocation = locations.sorted { $0.horizontalAccuracy < $1.horizontalAccuracy }.first
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        if status == .authorizedAlways || status == .authorizedWhenInUse {
-            locationManager.startUpdatingLocation()
-        } else {
-            locationManager.stopUpdatingLocation()
-        }
-        
-    }
-    
-}
